Split apart AutoCompleteBox into multiple files
